вынесла валидацию в отдельный модуль
